api/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
watchfiles==0.24.0
redis==5.0.8
pydantic-settings==2.2.1
